Add user accounts
Users can now register, log in, and log out. Registered accounts are
saved to a database. Usernames must be unique. Form validation is
currently only done on the backend and displayed as a SQL error. Logged
in users have a different nav. Currently, users who are not logged in
can still submit posts and topics (still frontend-only).